backend/requirements.txt
uvicorn
pymysql
python-dotenv
mysql-connector-python
flask